auth, cancel, costumer

+++ routes/api.php
@@ -4,6 +4,9 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\CarController;
 use App\Http\Controllers\OrderController;
+use App\Http\Controllers\AuthController;
+use App\Http\Controllers\CostumerController;
+use App\Http\Controllers\CancelController;
 
 /*
 |--------------------------------------------------------------------------
@@ -20,15 +23,39 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-// car routes
-Route::post('/car/{id}', [CarController::class, 'update']);
-Route::post('/car', [CarController::class, 'store']);
+Route::post('/register', [AuthController::class, 'register']);
+Route::post('/login', [AuthController::class, 'login']);
+
+// car router
 Route::get('/car/{id}', [CarController::class, 'show']);
 Route::get('/car', [CarController::class, 'index']);
-Route::delete('/car/{id}', [CarController::class, 'destroy']);
 
-// order routes
-Route::get('/order', [OrderController::class, 'index']);
-Route::post('/order/make_order', [OrderController::class, 'make_order']);
-Route::put('/order/start_rental/{id}', [OrderController::class, 'start_rental']);
-Route::post('/order/end_rental/{id}', [OrderController::class, 'end_rental']);
+// must login
+Route::middleware('auth:sanctum')->group(function(){
+    // == only admin ==
+    // car
+    Route::post('/car/{id}', [CarController::class, 'update'])->middleware('admin');
+    Route::post('/car', [CarController::class, 'store'])->middleware('admin');
+    Route::delete('/car/{id}', [CarController::class, 'destroy'])->middleware('admin');
+    // costumer
+    Route::get('/costumer/detail', [CostumerController::class, 'all_detail_user'])->middleware('admin');
+    Route::get('/costumer/detail/{id}', [CostumerController::class, 'detail_user'])->middleware('admin');
+    // order
+    Route::put('/order/start_rental/{id}', [OrderController::class, 'start_rental'])->middleware('admin');
+    Route::get('/order', [OrderController::class, 'index'])->middleware('admin');
+    // == all users ==
+    // order
+    Route::get('/order/{id}', [OrderController::class, 'show']);
+    Route::post('/order/make_order', [OrderController::class, 'make_order']);
+    Route::post('/order/end_rental/{id}', [OrderController::class, 'end_rental']); 
+    // cancel
+    Route::post('/cancel', [CancelController::class, 'cancel']);
+    Route::get('/cancel/{id}', [CancelController::class, 'reason']);
+    // costumer
+    Route::get('/costumer/myprofile', [CostumerController::class, 'my_profile']);
+    Route::post('/costumer', [CostumerController::class, 'add_costumer']);
+    Route::post('/costumer/update', [CostumerController::class, 'update']);
+    // auth
+    Route::post('/logout', [AuthController::class, 'logout']);
+    Route::get('/costumer', [CostumerController::class, 'add_costumer']);
+});
